api/requirements.txt
fastapi>=0.124.4
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0

# Dépendances de test
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
types-psycopg2==2.9.5
black>=25.12.0
ruff>=0.0.297
mypy>=1.3.0
pylint>=2.20.0